database updated with all 26 flag details, including images stored on firebase. however, images on firebase appear to be problematic for now. so ill fix that next. Data are comming through okay and ui looks good. database now allows access from all IP address. im not really okay with this
