+++ package.json
@@ -12,6 +12,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^0.25.0",
     "mongoose": "^6.2.0",
     "next": "12.0.8",
     "next-pwa": "^5.4.4",
