+++ pages/tutorial/index.js
@@ -1,11 +1,12 @@
 import Head from "next/head";
 import Link from "next/link";
+import axios from "axios";
 import BareMinimumTemplate from "../../layouts/interface/minimum";
 import styles from '../../styles/Tutorial.module.css';
 import SlidingElement from "../../components/slide";
 
 
-export default function Tutorial() {
+export default function Tutorial({flaglist}) {//flaglist is from the getServerProps below whish is like useEffect except it works before page loads
 
   return (
     
@@ -17,25 +18,34 @@ export default function Tutorial() {
       {/*This section should take the entire screen (consideration should be given to browser-views)*/}
 
             <div className={styles.slider}> {/*boostrap 'conatiner' class looks good but the width of the sliding element gets a bug. alittle portion of next element is visible*/}
-              <SlidingElement id="1" />
-              <SlidingElement id="2" />
-              <SlidingElement id="3" />
-              <SlidingElement id="4" />
+
+              {
+                flaglist.map((f)=>{
+                  return(
+                      <SlidingElement id={f._id} key={f._id} alphabet={f.name.charAt(0)}  name={f.name} meaning={f.meaning[0]} />
+                    )
+                })
+              }
+            {/*img_url={f.image}, firebase error i suppose*/}
+
             </div>
 
             <div className={styles["nav-group"]}>
-              <Link href="#slide-1" passHref> 
-                <a> A </a> 
-              </Link>
-              <Link href="#slide-2" passHref> 
-                <a> B </a> 
-              </Link>
-              <Link href="#slide-3" passHref> 
-                <a> C </a> 
-              </Link>
-              <Link href="#slide-4" passHref> 
-                <a> D </a> 
-              </Link>
+
+              {
+                flaglist.map((f)=>{
+                  
+                  return(
+
+                    <Link href={`#slide-${f._id}`} passHref key={f._id}> 
+                      <a> {f.name.charAt(0)} </a> 
+                    </Link>
+
+                  )
+                })
+              }
+
+              
             </div>
 
         </section>
@@ -56,4 +66,13 @@ export default function Tutorial() {
     
 
   )
+}
+
+export const getServerSideProps = async() =>{
+  const res = await axios.get("http://localhost:3000/api/v1/flags");
+  return{
+    props:{
+      flaglist: res.data,
+    }
+  }
 }
